Add base for restaurant page view
