+++ src/pages/RestaurantPageInfo/RestaurantPageInfo.scss
@@ -1,0 +1,13 @@
+.restaurant-picture {
+  width: 100%;
+  z-index: -1;
+}
+
+.contentt {
+  background-color: white;
+  padding: 20px;
+  border-radius: 20px;
+  border: 1px solid #BBBBBB;
+  position: relative;
+  top: -20px;
+}
